requirements.txt
pyannote.audio>=2.0
pyannote.core
pyannote.database
pyannote.metrics
tqdm
pandas
whisper
torch
